portx64: changed pointers stored in u32 to uintptr_t
- adjusted hardcoded mask/hardcoded offsets used in memory operations (modelasm_c.c)
- makefile: enabled x64 support and added lib dependency
- title: increase alloc size
- tex: adjustment to calculations involving GDL sizes

+++ src/game/bondgun.c
@@ -6489,9 +6489,9 @@ s32 bgunCreateModelCmdList(struct model *model, struct modelnode *nodearg, s32 *
 			rwdata->distance.visible = false;
 			node->child = rodata->distance.target;
 			ptr[0] = 0;
-			ptr[1] = (s32)rwdata;
-			ptr[2] = (s32)node;
-			ptr[3] = (s32)rodata->distance.target;
+			ptr[1] = (uintptr_t)rwdata;
+			ptr[2] = (uintptr_t)node;
+			ptr[3] = (uintptr_t)rodata->distance.target;
 			ptr += 4;
 			len += 16;
 			break;
@@ -6501,9 +6501,9 @@ s32 bgunCreateModelCmdList(struct model *model, struct modelnode *nodearg, s32 *
 			rwdata->toggle.visible = true;
 			node->child = rodata->toggle.target;
 			ptr[0] = 1;
-			ptr[1] = (s32)rwdata;
-			ptr[2] = (s32)node;
-			ptr[3] = (s32)rodata->toggle.target;
+			ptr[1] = (uintptr_t)rwdata;
+			ptr[2] = (uintptr_t)node;
+			ptr[3] = (uintptr_t)rodata->toggle.target;
 			ptr += 4;
 			len += 16;
 			break;
@@ -6512,7 +6512,7 @@ s32 bgunCreateModelCmdList(struct model *model, struct modelnode *nodearg, s32 *
 			rwdata->headspot.headmodeldef = NULL;
 			rwdata->headspot.rwdatas = NULL;
 			ptr[0] = 2;
-			ptr[1] = (s32)rwdata;
+			ptr[1] = (uintptr_t)rwdata;
 			ptr += 2;
 			len += 8;
 			break;
@@ -6520,7 +6520,7 @@ s32 bgunCreateModelCmdList(struct model *model, struct modelnode *nodearg, s32 *
 			rwdata = modelGetNodeRwData(model, node);
 			rwdata->type0b.unk00 = 0;
 			ptr[0] = 3;
-			ptr[1] = (s32)rwdata;
+			ptr[1] = (uintptr_t)rwdata;
 			ptr += 2;
 			len += 8;
 			break;
@@ -6528,7 +6528,7 @@ s32 bgunCreateModelCmdList(struct model *model, struct modelnode *nodearg, s32 *
 			rwdata = modelGetNodeRwData(model, node);
 			rwdata->chrgunfire.visible = false;
 			ptr[0] = 4;
-			ptr[1] = (s32)rwdata;
+			ptr[1] = (uintptr_t)rwdata;
 			ptr += 2;
 			len += 8;
 			break;
@@ -6539,10 +6539,10 @@ s32 bgunCreateModelCmdList(struct model *model, struct modelnode *nodearg, s32 *
 			rwdata->dl.gdl = rodata->dl.opagdl;
 			rwdata->dl.colours = (void *)ALIGN8((uintptr_t)&rodata->dl.vertices[rodata->dl.numvertices]);
 			ptr[0] = 5;
-			ptr[1] = (s32)rwdata;
-			ptr[2] = (s32)rwdata->dl.vertices;
-			ptr[3] = (s32)rwdata->dl.gdl;
-			ptr[4] = (s32)rwdata->dl.colours;
+			ptr[1] = (uintptr_t)rwdata;
+			ptr[2] = (uintptr_t)rwdata->dl.vertices;
+			ptr[3] = (uintptr_t)rwdata->dl.gdl;
+			ptr[4] = (uintptr_t)rwdata->dl.colours;
 			ptr += 5;
 			len += 20;
 			break;
